database_agent/requirements.txt
oracledb>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
hello-agents>=0.1.0
tabulate>=0.9.0